+++ server/src/main/kotlin/org/kryptonmc/krypton/world/chunk/ChunkSection.kt
@@ -1,19 +1,25 @@
 package org.kryptonmc.krypton.world.chunk
 
 import org.kryptonmc.krypton.api.registry.NamespacedKey
-import org.kryptonmc.krypton.registry.Registries
-import org.kryptonmc.krypton.world.block.palette.GlobalPalette
-import org.kryptonmc.krypton.world.data.PalettedContainer
+import org.kryptonmc.krypton.api.space.Vector
+import org.kryptonmc.krypton.world.block.KryptonBlock
 import org.kryptonmc.krypton.world.data.StateIndexHolder
 
 data class ChunkSection(
     val y: Int,
-    private var nonEmptyBlockCount: Int = 0,
     val blockLight: ByteArray,
     val skyLight: ByteArray,
+    val palette: MutableList<ChunkBlock>,
+    val blockStates: StateIndexHolder
 ) {
 
-    private val states = PalettedContainer(GLOBAL_BLOCK_STATE_PALETTE, Registries.BLOCK_STATES, AIR)
+    fun setBlock(location: Vector, block: KryptonBlock) {
+        val index = indexOf(location.x.toInt() and 0xF, location.y.toInt() and 0xF, location.z.toInt() and 0xF)
+        val state = palette.indexOfFirst { it.name == block.type.key }
+        blockStates[index] = state
+    }
+
+    private fun indexOf(x: Int, y: Int, z: Int) = (y shl 8) or (z shl 4) or x
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
@@ -23,12 +29,6 @@ data class ChunkSection(
     }
 
     override fun hashCode() = arrayOf(y, blockLight, skyLight, palette, blockStates).contentDeepHashCode()
-
-    companion object {
-
-        private val AIR = NamespacedKey(value = "air")
-        private val GLOBAL_BLOCK_STATE_PALETTE = GlobalPalette(Registries.BLOCK_STATES, AIR)
-    }
 }
 
 data class ChunkBlock(